1. This submission is the MTR test in mysql-test of parallel query. 2. Here are the parameters and Status variables of the parallel query:
============parameters============
============parallel_cost_threshold============
Global- and session-level parameter, which is used to set the threshold for parallel query of SQL statements.

Parallel queries are executed only when the estimated cost of a query is higher than this threshold. When the estimated cost is lower than this threshold, a native query process is executed.

Value range: 0 to ULONG_MAX
Default value: 1000

============parallel_default_dop============
Global- and session-level parameter, which is used to set the maximum number of parallel queries for each SQL statement.

The query parallelism of a SQL statement is dynamically adjusted based on the table size. If the binary tree of the table is too small (the number of table slices is less than the concurrency), set the query concurrency based on the number of table slices. The maximum degree of parallelism for each query does not exceed the value of parallel_default_dop. The value of this parameter cannot be greater than that of parallel_max_threads. Otherwise, parallel query of SQL statements cannot be enabled.

Value range: 0 to 1024
Default: 4

============parallel_max_threads============
Global-level parameter, which is used to set the total number of concurrent query threads.

Value range: 0 to ULONG_MAX
Default value: 64

============parallel_memory_limit============
Global-level parameter, which is used to set the upper limit of the total memory used by the leader thread and worker threads in parallel execution.

Value range: 0 to ULONG_MAX
Default value: 100*1024*1024

============parallel_queue_timeout============
Global- and session-level parameter, which is used to set the timeout interval for parallel queries.

If system resources are insufficient, for example, the number of running parallel query threads reaches the value of parallel_max_threads, the parallel query statements will wait for available resources. If no resource is available after timeout, native query is executed.

Value range: 0 to ULONG_MAX, in milliseconds
Default value: 0

============force_parallel_execute============
Global- and session-level parameter, which is used to set the switch for parallel queries.

The bool value is on or off.
The value on enables parallel query while off disables parallel query.

Default value: off

============Status Variable============
============PQ_threads_running============
Global-level variable, which indicates the total number of running parallel threads.

============PQ_memory_used============
Global-level variable, which indicates the total memory used for parallel execution.

============PQ_threads_refused============
Global-level variable, which indicates the total number of parallel queries that cannot be executed due to the limit on the total number of threads.

============PQ_memory_refused============
Global-level variable, which indicates the total number of parallel queries that fail to be executed due to the total memory limit.

3.Possible Incompatibility with Serial Results

The parallel execution result may be incompatible with the serial execution result in the following scenarios:
3.1.The number of errors or alarms may increase.
For queries that encounter errors or alarms during serial execution, if they are executed in parallel, each worker thread may report errors or alarms. As a result, the total number of errors or alarms increases.

3.2.The precision is incorrect.
During parallel execution, intermediate results may be generated. If the intermediate results are of the floating-point type, the floating-point data may have precision deviation. As a result, the final result may be slightly different.

3.3.Sequences of the result sets are different.
When multiple worker threads execute the query, the returned result set may have a different sequence from the serial execution sequence. If the GROUP BY statement exists, the sequence in the group after grouping may be different from the serial execution sequence. If the LIMIT statement exists, it is more likely that the parallel execution result is different from the serial execution result.

4.Pictures of the principle of parallel query implementation and the process of key functions have been added to the pq_doc folder of the project. I think it will be helpful for subsequent development.